ecommerce/requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==4.2.15
django-appconf==1.0.6
django-compressor==4.5.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
Flask==3.0.3
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
pillow==10.4.0
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.20.0
